requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
pymongo>=4.4
python-dotenv>=1.0
python-multipart>=0.0.6
Pillow>=9.0
pytest>=7.0
